Prueba de Test | Problemas con base de datos FK (Solucionado)

+++ tienda_informatica_final/src/main/java/com/tienda/model/Telefono.java
@@ -2,25 +2,25 @@ package com.tienda.model;
 
 public class Telefono {
     private int idTelefono; 
-    private int telefono;
+    private long telefono;
     private String tipoTelefono;
+    
 
     public Telefono() {
     }
     
 
-    public Telefono(int telefono, String tipoTelefono) {
+    public Telefono(long telefono, String tipoTelefono) {
         this.telefono = telefono;
         this.tipoTelefono = tipoTelefono;
     }
 
 
-    public Telefono(int idTelefono, int telefono, String tipoTelefono) {
+    public Telefono(int idTelefono, long telefono, String tipoTelefono) {
         this.idTelefono = idTelefono;
         this.telefono = telefono;
         this.tipoTelefono = tipoTelefono;
     }
-    
     // Getters y Setters
     public int getIdTelefono() {
         return idTelefono;
@@ -30,11 +30,11 @@ public class Telefono {
         this.idTelefono = idTelefono;
     }
 
-    public int getTelefono() {
+    public long getTelefono() {
         return telefono;
     }
 
-    public void setTelefono(int telefono) {
+    public void setTelefono(long telefono) {
         this.telefono = telefono;
     }
